Send secret header to API from CloudFront (#1186)
This is part 2 of #1181, and is a follow-up to #1185. It creates a secret and has CloudFront send it as a header when proxying requests to the API service. (It does not yet make the API load balancer require the secret, since the CloudFront distribution takes a while to change, and we don't want to break the service while that's happening.)

This also fixes a variable description that it turns out I got wrong in #1185.