Only allow requests through CloudFront, take 2 (#1188)
This locks down the API server's load balancer so that it only forwards requests to the actual service if it has a header with a secret value. CloudFront is already configured to set that header (see #1186), so this effectively allows only requests from CloudFront.

I tried to do this in #1186 with a wildcard host header, but it turns out AWS rejects that once you actually try and apply it. This approach uses an IP match to allow all IP addresses instead, which I tested in AWS already so it should work.

This also doesn’t set the default rule to block responses, so if this goes wrong, things still work fine.