Pass auth headers through CloudFlare (#1191)
We were accidentally swallowing auth-related headers in CloudFlare! We probably never noticed because the loaders (and other data-submitting partners in the past) went direct to the load balancer. (Noticed in the mess with #1190.) At any rate, this resolves the issue, so we could have the loaders talk through CloudFlare now if we wanted.