Support bivalent infant vaccines at Hyvee (#1202)
HyVee recently started using the vaccine code `Pediatric Moderna Bivalent 3-5`, and this adds support for it. It also adds support for the code `Pediatric Moderna Bivalent (3-5)` (note the parentheses) even though we haven’t seen it, since that matches the naming convention for all the other codes so far — I expect the one we are seeing is likely to change to this in the near future.

I added support for the matching Pfizer vaccine while I was at it, even though we haven’t seen it yet either (the name based on the existing conventions we’ve seen so far).

Fixes https://sentry.io/organizations/usdr/issues/3828195366.