Add retry support for HyVee
We've had a lot of random HyVee failures that are quite noisy, and I'm hoping retrying a few time son failure will resolve the issue. We appear to have always disabled retry for HyVee (see #436), but with no note about why. I wrote it and I don't remember, so I'm thinking this was just an oversight from testing that didn't get corrected before merging.