Handle renamed fields in HyVee API, add retries (#1266)
HyVee renamed the `manufacturerName` field to `vaccineName`. The values appear to be the same; this seems to just be a long-coming acknowledgement of the fact that this field isn't really about *manufacturers*.

While I was at it, I added retry support for HyVee (we retry on most sources). We've had a lot of random HyVee failures that are quite noisy, and I'm hoping retrying a few time son failure will resolve the issue. Retry has always been disabled for HyVee (#436), but with no note about why. I wrote it and I don't remember, so I'm thinking this was just an oversight from testing that didn't get corrected before merging.