Turn on versioning for data snapshots bucket
At the end of last week, I discovered some bugs with the writing of historical data to out S3 buckets, and while some of those issues are "fixable", others are not because the original data was lost. Having bucket versioning turned on would have allowed us to resolve those issues.