Add better retry support for GraphQL
I tried to add retry support for Hy-Vee in #1266. It reduced HTTP errors slightly, but we're still seeing a bunch in production. After more noodling, I think the API is returning a 200 status code when it was able to catch and format the errors for GraphQL (the GraphQL spec does not define what HTTP status codes should be used!) so our HTTP client's built-in retry logic doesn't do anything, and no retries occur. This adds a new `queryGraphQl()` helper that gives a way to handle that (and also provides some other useful defaults).