Compress data sent from the loaders (#1321)
The loaders can send pretty significant amounts of data to the API server, so there's potentially a lot to be gained from compressing it. We already support compressed POST bodies on the server, so all that needs to happen here is telling the loader's client to compress when sending. Unfortunately Got only has built-in support for compressed responses, not requests, so this is a little bit manual.