Run loaders on UNIVAF public subnets
This moves our loaders back onto the UNIVAF VPC, but using the public subnet instead of the private one. In #1322 and #1323, I misunderstood the Internet Gateway vs. NAT Gateway that were operating in our various subnets. What we really needed was for loaders to have public IPs (they always have, it turns out!) and an IG. Our private subnets don't have an IG, and our public networks do (that's *mainly* what makes them private vs. public). The default subnets all have IGs. So really we just need to run the loaders in the UNIVAF VPC's public subnet. The other lesson learned here was that the smart thing in this case is probably to also give them a new Security Group that has no ingress at all (so they can make connections to other services, but not the other way around -- important since they are otherwise publicly addressable because of the public IP and IG).