Change loader security group VPC (correctly) (#1325)
In #1324, I tried to change the loader security group's VPC, but it turns out that under the hood you can't do that -- Terraform will try to destroy and recreate it. However, things get messy since there are always loader jobs running, and it can't actually do the destroy and recreate while balancing all the dependencies. This gives the security group a new name, which should solve the issue.