Move all ECS tasks to public network
This is the first step of #1327. It moves the API server tasks and the daily data snapshot task from our private subnets to our public subnets (we already did this for the loaders). It also adds a security group for cron jobs that is very strict (the daily data snapshot task was using a security group designed for a server, which is not ideal) and can be used by all our cron-job-like tasks. It also creates a new security group for the API service that is more appropriately named for clarity (the old name may be how we wound up using it for the daily data snapshot).

A follow-up change will remove the old security groups once everything has transitioned over to the new ones.