Flip dependency relationship on DB security
The DB is currently set up to only accept ingress from resources with a given set of security groups. That makes setup and dependency management a bit backward, where the security groups for other resources that need the database have to be created before the database, which needs to be created before the resources that need the security groups, which leads to those security groups being placed away from the rest of the configuration for those services. There are some messy implicit dependencies.

This flips things around by having the database module create a security for itself and a security group for other resources that want to access it. It outputs the ARN & ID for the second security group, and then services that want to talk to the database should use those outputs. This makes the implicit dependency on the database explicit (the API service, for example, needs access to the database, and by reading the security group it needs from the database, creates an actual dependency on the DB in the Terraform world).

Since that's cleaned up, this then moves all the security group definitions alongside the services/resources that they apply to.

NOTE THIS NEEDS FOLLOW-UP to remove the database's dependency on the security groups for the services that use the database, since those will no longer be needed once this change is fully deployed. (That has can't be part of this change, though, since we can't remove access for those security groups until new services are deployed with the new database access security group.)