Reduce concurrency for CVS, Walgreens (#1369)
A couple weeks ago I shrunk the API servers and that caused them to start dropping requests (totally manageable; the loaders retry and never experienced any failures as a result). Reducing the concurrency appears to help alleviate the issue without adding more resources (and it's not actually taking much longer to run the jobs).