Repair the Rite Aid scraper (#1370)
Our Rite Aid scraper broke a few days ago, and it seems like Rite Aid made some small but complicated changes to their web scheduler’s API. This fixes things for New Jersey and a few other states, but leaves some other ones that are complicated and low priority broken. It’ll need a second pass.

The main thing is that you can no longer request more than 100 locations at a time. There is no pagination mechanism, so I needed to re-think the zip codes and radii we query with in order to make sure we actually get all the results. Even worse, I realized partway through that the API is not actually returning all the locations in the given radius (!). I spent way too much time trying to figure out what’s going on, and I don’t get it. I think the actual effective radius you get might have something to do with the density of stores in a given area, but I really can’t understand what it’s doing from the outside. That meant a lot of time playing with different configurations of zipcodes and radii.

As a result, I built a working zipcode list for New Jersey that seems to return all the stores, but for the other states, I checked the number of stores it was getting against the [store list on riteaid.com](https://www.riteaid.com/locations/index.html) and left it alone if it got everything or added a `FIXME` comment and removed it from the `--states` option in the cron job configuration if not. I’ll have to come back to those later.

There were some other minor changes:
- Rite Aid no longer has any stores in Colorado (this has been true for more than a year, but querying in an empty state only started causing problems now).
- The format of some fields changed.
- Added bivalent booster support.

Fixes https://usdr.sentry.io/issues/2420647478