Stop disabling Sentry during tests
This used to be needed in order for all the tracing APIs we use to work right during tests, but now it's breaking them. It seems like subsequent `Sentry.init()` calls are causing fresh copies of the Sentry modules to be loaded, so they are no longer using the classes we monkey-patch. In *theory*, turning off Sentry sending during tests like this is the right thing to do, but it looks complicated to do and keep things working now.