Apply rate limit to all Rite Aid scraper requests
The rate limit for the Rite Aid scraper has been erroneously applied just to the initial "summary data" requests, and not to the secondary "detailed" requests for slot-level info. That's a mistake! We are getting a few 403 (forbidden) responses to the scraper's requests every hour, and remembering to apply the rate limit to all requests is likely to help here.