Retry Rite Aid requests on 403 status codes
The Rite Aid API and scraper both occasionally encounter a raft of 403 status codes (anywhere from 1/hour to 1/day) and fail. This sets us up to retry those requests rather than failing immediately, but with a greater than usual delay under the assumption that this is a rate-limiting/abuse measure.

This is an experiment! It might not fix the issue -- an equally likely possibility is that because we run on Fargate, each loader run comes from one of a wide set of IP addresses, and some of them have been blocked by Rite Aid. If that's the case, retrying won't do us any good.