Fix Rite Aid 403 retry exception (#1398)
I made a mistake in #1397 by assuming every time the `calculateDelay` function runs for retries, the error involved an actual HTTP response. Some are network problems where the client is unable to communicate with the server at all. (I tested 403 responses, but did not test this!)

This also consolidates the logic in one spot, since it was obviously more complicated than I gave it credit for (I didn’t bother before since it was experimental. That was wrong.).

Fixes https://usdr.sentry.io/issues/4048563618