Rite Aid API: stop early on auth errors
We used to keep trying other states in the Rite Aid API after getting auth errors because the API does not differentiate between an auth error and a request for a state with no Rite Aid stores in it (!). However, some of the IPs we wind up running on with Fargate are now blocked by Rite Aid, and this leads to us pointlessly trying HTTP requests for every state, even when we know they are blocked. Since the list of valid states is pretty predictable, stop trying additional states in this situation and just bail out early. This also allows us to use more straightforward rate-based ignore rules in Sentry so we aren't bothered by this situation.