Clean up Rite Aid auth error handling (#1410)
This reverts #1397 and ensures we bail out early in the Rite Aid API if we hit auth errors (we already do so in the scraper). We previously kept trying other states on auth errors because requests for states with no stores would return 401 or 403 status codes, and we wanted to be robust against that. However, Rite Aid appears to have started blocking some of the IP addresses for the Fargate machines we run on, so we occasionally get all our requests blocked with auth errors. (This is obviously a way bigger problem than us winding up querying a state with no stores because Rite Aid closed them all, which is what happened before.)

This approach lets us stop early and better manage rate-based ignore rules in Sentry (we don’t get blocked on most runs, so as long as the rate of failing runs doesn’t get too high, it’s fine).

While I was at it, I noticed some of the errors we get from the Rite Aid *Scraper* for this case are misclassified in Sentry. It turns out that Sentry infers an error’s type based on its `name` property rather than its actual type (i.e. by checking `error.constructor.name`). I’ve added a base class for all errors that sets the `name` property based on the class name. This also fixes an oddball inconsistency between `console.log(error)` and `error.toString()`.