Rite Aid: Scrape less, use API more
We've had the Rite Aid Scraper running every 10 minutes and the API source running every 30 for a long time, largely owing to the fact that the main API was lacking detailed information we wanted from the scraper, and that the scraper was not very slow compared to the API. These days, we no longer need as detailed or up-to-date info and the scraper has become much slower (see #1370, #1371, #1396). The scraper is expensive to run for both us and Rite Aid, and it's time to turn it down a bit.