Lower API concurrency for for all loader jobs
A while back, I lowered the API concurrency for CVS and Walgreens because they were causing a lot of errors and retries. That made the issue occur a lot less and surprisingly did not have a notable impact on loader job run time! It probably makes sense to apply across all jobs.