Rebalance rite aid loaders (#1411)
We've had the Rite Aid Scraper running every 10 minutes and the API source running every 30 for a long time, largely owing to the fact that the main API was lacking detailed information we wanted from the scraper, and because the scraper was actually reasonably speedy compared to the API. These days, we no longer need as detailed or up-to-date info and the scraper has become much slower (see #1370, #1371, #1396). The scraper is expensive to run for both us and Rite Aid, so I’m turning down to every 30 minutes (and running the API slightly more often).

While I was at it, I also lowered the API concurrency for all loader jobs. I did this a while ago just for CVS and Walgreens to try and reduce retries and other internal errors. It worked, and did not even make the loaders any slower like I’d expected. Based on that, it probably makes sense to turn concurrency down for *all* loader jobs.