Remove duplicate vaccine types/constants
Way back in #1133, I "moved" all the types and constants describing vaccines (CVX codes, product names, ages, etc.) to univaf-common. However, it seems I forgot to remove the old version that was in the server package. Some references got updated and others didn't, leading to subtle bugs as the two modules diverged. This resolves the issue by deleting the version of the module that is in the server package and ensuring all references point to the more up-to-date one in univaf-common.