Stop filtering locations based on old data file
The `known-njvss-locations.json` data file is unmaintained and no longer relevant. This drops filtering based on it, and also removes a lot of spurious debug logging we haven't been paying attention to for a long time. In practical terms, this doesn't appear to change behavior since we were not actually filtering any locations out as a result of this anyway.