Stop manually matching location IDs
This is no longer necessary (and hasn't been for a long time!). We have enough in the `external_ids` field for the server to reliably match NJVSS records to existing locations in the database. I also tested this against the real data we currently have in the DB and its behavior does not differ in practice.