Disable current Albertsons source
The Alberstons loader source reads data from MHealth, but Albertsons has finally switched from MHealth to their own booking system for COVID vaccines, and all the data in the API we were querying is now stale and irrelevant. We probably should write a new scraper (see #1460), but in the mean time, we should disable the existing one.