Keep track of stale data in loader
On each loader run, send metrics about stale data to Datadog and log warnings. This also adds a command-line option to filter out stale data from the result set, too. A record's age is based on the older of `availability.valid_at` and the latest `availability.capacity` entry or latest `availability.slots` entry (sometimes `valid_at` turns out to be a lie, so we use the actual slots or capacity when we can).

The two added CLI options:
- `--filter-stale-data` will cause the loader not to output/send location records that are stale.
- `--stale-threshold N` sets the threshold (in milliseconds) at which stale data is filtered out or warnings are logged. (Metrics are sent to Datadog regardless of the threshold.)

In all cases, metrics are sent to Datadog (as long as Datadog is configured), matching the fields used for histograms:
- `loader.loader.data.age_seconds.min`
- `loader.loader.data.age_seconds.max`
- `loader.loader.data.age_seconds.avg`
- `loader.loader.data.age_seconds.median`

Fix #1459.