Turn CVS and Kroger schedules WAY down (#1496)
The scuttlebut from a Kroger developer who is no longer working on it is that the team intentionally decided to turn it off (without notifying any government partners), so we probably shouldn't expect it to come back at this point. CVS's data is also so old I'm sure we're seeing something similar. This turns down the schedules for them to only run 4 times a day, since we don't expect to get meaningful data. At this point, they're just there to notify us if they somehow stop failing.