Add store list scraper for CVS
This is pretty analagous to the scraper we wrote a while back to get all Albertsons stores. I wrote this during the mess around #1458 and have had it just sitting uncommitted for a while. In the end, it wasn't working more on since we are ramping things down, but it would probably be useful for anyone who might want to make use of the scrapers and tools in this codebase, so I felt it was worth committing for posterity and archival purposes.