Clean up loader logging
Another not-totally-necessary-but-I've-wanted-to-do-it-for-a-while-and-it's-cathartic cleanup refactor in the loader: this moves logging functionality to its own module, creates a more organized `Logger` class that handles common tasks across different outputs and levels, etc.

This simplifies a bunch of logging code in little tiny ways. The main thing is that consuming code no longer chooses whether to send contextual data to Sentry. It just gets sent (in practice, we have always sent the data since we added that functionality; the extra argument for it is no longer meaningfully necessary).

Fixes #306. This doesn't use a third-party library like Winston, but solves most of the small issues that one needed to serve, with less last-minute complexity that switching to a new logging library (if this project weren't getting archived in several weeks, this might be a different story).