Unnecessary but cathartic refactor 2: loader logging (#1525)
Another not-totally-necessary-but-I've-wanted-to-do-it-for-a-while-and-it's-cathartic cleanup refactor in the loader: this moves logging functionality to its own module, creates a more organized `Logger` class that handles common tasks across different outputs and levels, etc.

This simplifies a bunch of logging code in little tiny ways and also cleans up a bunch of spots where we used `console.warn/error()` directly because we didn’t want to sent something Sentry (now you can just `logger.info()` or `logger.debug()`).Consuming code also longer chooses whether to send contextual data to Sentry. It just gets sent since, in practice, we have always sent the data after we added “context” functionality to the Sentry logic.

It also effectively fixes #306. It doesn't use a third-party library like originally intended, but solves most of the small issues that were important without the complexity of a totally new logging system (if this project weren't getting archived in several weeks, this might be a different story).

Finally, this stops sending the warning to Sentry about a source not having any info about staleness, since that’s not actually useful in practice, and just generates lots of noise we have to ignore.