HOTFIX: Support nested arrays in Walgreens
Walgreens's SMART SL manifest file currently lists the slots as a nested array of output items inside the main `output` array, which is not correct, and has been causing us to list all Walgreens as unavailable. This hacks in a fix (we just flatten the `output` array before handling it). I've also alerted the maintainer of the Walgreens API to the issue, so hopefully the underlying system gets fixed soon.