HOTFIX: Support nested arrays in Walgreens (#1544)
Walgreens's SMART SL manifest file currently lists the slots as a nested array of output items inside the main `output` array, which is not correct, and has been causing us to list all Walgreens as unavailable.

That is, the manifest looks like:

    {
      "transactionTime": "2023-06-06T03:21:31.360Z",
      "request": "...",
      "output": [
        { "type": "Location", "url": "..." },
        { "type": "Schedule", "url": "..." },
        
        // These items should not be nested in a sub-array!
        [
          { "type": "Slot", "url": "...", "extension": { "state": ["AK"] } },
          { "type": "Slot", "url": "...", "extension": { "state": ["AL"] } },
          // etc...
        ]
      ]
    }

This hacks in a quick fix: we just flatten the `output` array before handling it. I've also alerted the maintainer of the Walgreens API to the issue, so hopefully the underlying system gets fixed soon.