Valid schema for SMART SL data
We recently had the Walgreens SMART Scheduling Links API break (#1544), and only found out about it because I happened to be looking into some unrelated errors that I thought were happening with Walgreens (they were not) and noticed some fishy output. That's not good, so this adds schema validation to the SMART Scheduling Links utilities, which would alert us to this issue if it happened again.

This also removes yesterday's hotfix for the Walgreens issue, since it already got fixed at the source! It's nice when some providers are quick and attentive. :)

Fixes #1545.