Archive a clean copy of Terraform configs
For the rest of this week, I'll be planning and implementing a shutdown of our production services, which means making changes to Terraform that someone forking and re-deploying won't want to use. So others have our normal production configurations as a guide, I've made a copy of all the current files in `terraform/deprecated/`.

This is part of #1550.