Add note about AWS Terraform provider version
We use a Terraform module that isn't yet compatible with v5 of the AWS provider, and this adds a note about the situation, since it might not be clear to someone coming in fresh to the codebase. Hopefully Cloudposse will update the module we use, but it doesn't look like that's coming soon, given that they recently updated their tutorials to specify AWS < 5. (See https://github.com/cloudposse/tutorials/pull/21#discussion_r1219948021)