More yield* !
Previously, I used `yield * iterable` in some places, but in others I left a loop through an iterable that just yielded each entry. That was during a time when I was trying to reconcile how to best handle the difference between sources that needed to send additional options with each yield and those that didn't. Later I just let some sources return location records that weren't wrapped in an array, which made `yield*` possible, but didn't fix all the callsites where it would have been applicable. This does so, and makes things consistent.