Absurdly unnecessary cleanup refactor: Make loader sources async iterators (#1563)
This fixes #55 by getting rid of the callback argument for `checkAvailability()` entirely. A source's `checkAvailability()` function should now be an async generator that yields either: location records or arrays of `[locationRecord, sendOptions]`.

One of the weird and messy things in the loader sources has been that their main function (`checkAvailability()`) takes a callback to run for every found location (so we get results outputting progressively and reducing load to the server receiving them) but also returns an array of all the found locations (mostly for easy testing). IIRC, I was originally going to make it return a stream instead, but the mechanics of that turned out to be pretty complicated. In #55 I wanted to at least clean it up a bit and position the callback in a more conventional spot, but I actually think async iteration makes the most sense here; it gets us the stream-like stuff with simpler, built-in syntax.

Not everything is as clean as I was hoping here; I had been thinking test code would be really simple because we could polyfill the new `Array.fromAsyc()` or `AsyncIterator.prototoype.toArray()` functions, but forgot that the callbacks optionally take a second argument for options about how the location is sent to the database, and so an array of a source's outputs is not necessarily an array of the location objects (that is, you can't just do `await checkAvailability().toArray()` and expect an array of locations). So the tests have a new helper to get just the locations.

The CLI module has slightly more code to manage this, too, but I hope the overall working of it is slightly clearer, since we can just compose a series of iterators using `stream.compose()` for all the result processing steps.