Add tombstone page (#1569)
This adds a really simple, static “tombstone” site that will replace the API server at https://getmyvax.org this evening when we shut things down. It’s intended to deploy on GitHub pages (replacing the current “demo UI”); we’ll just update the getmyvax.org DNS to point to it later on.

Note this will need a follow-on change to fill in some of the statistics once they are calculated.