Shutdown: Delete database and bastion (#1582)
We've archived final dumps of the database in S3 at archives.getmyvax.org, and all the services and resources that touch the database been been removed. It's time to drop the database, too. That leaves us without any reason for having the bastion server, so this also removes it.

Part of #1550.